requirements.txt
pymongo
pymongo[encryption]
pygments
python-dotenv
colorama